requirements.txt
# Backend Dependencies
Django==5.0.2
python-dotenv==1.0.0
mysqlclient==2.2.4  # Para la conexión con MySQL
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pytest==8.0.0  # Para pruebas 
pytest-django==4.8.0  # Para pruebas de Django
gunicorn==21.2.0  # Para servir la aplicación
whitenoise==6.6.0  # Para servir archivos estáticos

# Frontend Dependencies (Node.js)
## Production Dependencies
@ant-design/icons@5.4.0
@testing-library/user-event@14.5.2
antd@5.20.4
axios@1.7.7
bootstrap@5.3.3
jwt-decode@4.0.0
react@18.3.1
react-bootstrap@2.10.4
react-dom@18.3.1
react-icons@5.3.0
react-router-dom@6.26.1
react-scripts@5.0.1
sass@1.77.8
web-vitals@4.2.3


## Development Dependencies
@babel/plugin-proposal-private-property-in-object@7.21.11
@babel/preset-env@7.25.4
@babel/preset-react@7.24.7
@testing-library/jest-dom@6.5.0
@testing-library/react@16.0.1
babel-jest@29.7.0
jest@29.7.0
